requirements.txt
# Core Flask dependencies
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Admin>=1.6.0
Flask-Babel>=3.0.0
Flask-CORS>=4.0.0
Flask-Mail>=0.9.1
Flask-Migrate>=4.0.0

# Database
SQLAlchemy>=2.0.0

# Security and authentication
Flask-Login>=0.6.0
Werkzeug>=2.3.0

# Text processing (removed markdown - no longer needed)

# HTTP requests for testing
requests>=2.31.0

# Environment management
python-dotenv>=1.0.0

# Communications
twilio>=8.0.0
APScheduler>=3.10.0
pytz>=2023.3

# OpenAI API (if needed)
openai>=1.0.0

# Data processing
pandas>=1.5.0
requests>=2.28.0

# HTML to PNG conversion
selenium>=4.15.0
Pillow>=10.0.0
webdriver-manager>=4.0.0
imgkit>=1.2.0  # Alternative HTML to PNG converter (requires wkhtmltopdf)
unqlite
